backend/requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flatbuffers==1.12
flex==6.14.1
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpointer==2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
Pillow==8.2.0
protobuf==3.17.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
validate-email==1.3
Werkzeug==2.0.1
wrapt==1.12.1
